+++ relay-server/src/services/processor/span.rs
@@ -2,8 +2,6 @@
 
 use prost::Message;
 use relay_dynamic_config::Feature;
-use relay_event_normalization::span::tag_extraction;
-use relay_event_schema::protocol::{Event, Span};
 use relay_protocol::Annotated;
 use relay_quotas::DataCategory;
 use relay_spans::otel_trace::TracesData;
@@ -127,22 +125,6 @@ fn parse_traces_data(item: Item) -> Result<TracesData, DiscardReason> {
     }
 }
 
-/// Creates a span from the transaction and applies tag extraction on it.
-///
-/// Returns `None` when [`tag_extraction::extract_span_tags`] clears the span, which it shouldn't.
-pub fn extract_transaction_span(
-    event: &Event,
-    max_tag_value_size: usize,
-    span_allowed_hosts: &[String],
-) -> Option<Span> {
-    let mut spans = [Span::from(event).into()];
-
-    tag_extraction::extract_span_tags(event, &mut spans, max_tag_value_size, span_allowed_hosts);
-    tag_extraction::extract_segment_span_tags(event, &mut spans);
-
-    spans.into_iter().next().and_then(Annotated::into_value)
-}
-
 #[cfg(test)]
 mod tests {
 
@@ -246,6 +228,7 @@ mod tests {
             "instrumentation.scope_key": "scope_value",
             "instrumentation.version": "0.0.1",
             "resource.resource_key": "resource_value",
+            "sentry.origin": "auto.otlp.spans",
             "span_key": "span_value"
           },
           "exclusive_time": 0.0,