+++ tests/integration/test_attachments.py
@@ -135,13 +135,17 @@ def test_attachments_ratelimit(
     # First attachment returns 200 but is rate limited in processing
     relay.send_attachments(project_id, event_id, attachments)
 
-    outcomes_consumer.assert_rate_limited("static_disabled_quota")
+    outcomes_consumer.assert_rate_limited(
+        "static_disabled_quota", categories=["attachment", "attachment_item"]
+    )
 
     # Second attachment returns 429 in endpoint
     with pytest.raises(HTTPError) as excinfo:
         relay.send_attachments(project_id, event_id, attachments)
     assert excinfo.value.response.status_code == 429
-    outcomes_consumer.assert_rate_limited("static_disabled_quota")
+    outcomes_consumer.assert_rate_limited(
+        "static_disabled_quota", categories=["attachment", "attachment_item"]
+    )
 
 
 def test_attachments_pii(mini_sentry, relay):
@@ -380,7 +384,8 @@ def test_attachments_quotas(
     relay.send_attachments(project_id, event_id, attachments)
 
     outcomes_consumer.assert_rate_limited(
-        "attachments_exceeded", quantity=len(attachment_body)
+        "attachments_exceeded",
+        categories={"attachment": len(attachment_body), "attachment_item": 1},
     )
 
     # Second attachment returns 429 in endpoint
@@ -388,7 +393,8 @@ def test_attachments_quotas(
         relay.send_attachments(42, event_id, attachments)
     assert excinfo.value.response.status_code == 429
     outcomes_consumer.assert_rate_limited(
-        "attachments_exceeded", quantity=len(attachment_body)
+        "attachments_exceeded",
+        categories={"attachment": len(attachment_body), "attachment_item": 1},
     )
 
 
@@ -430,7 +436,8 @@ def test_attachments_quotas_items(
     relay.send_attachments(project_id, event_id, attachments)
 
     outcomes_consumer.assert_rate_limited(
-        "attachments_exceeded", categories=["attachment_item"], quantity=1
+        "attachments_exceeded",
+        categories={"attachment": len(attachment_body), "attachment_item": 1},
     )
 
     # Second attachment returns 429 in endpoint
@@ -439,7 +446,8 @@ def test_attachments_quotas_items(
     assert excinfo.value.response.status_code == 429
 
     outcomes_consumer.assert_rate_limited(
-        "attachments_exceeded", categories=["attachment_item"], quantity=1
+        "attachments_exceeded",
+        categories={"attachment": len(attachment_body), "attachment_item": 1},
     )
 
 
@@ -496,7 +504,6 @@ def test_event_with_attachment(
     relay_with_processing,
     attachments_consumer,
     transactions_consumer,
-    outcomes_consumer,
 ):
     project_id = 42
     event_id = "515539018c9b4260a6f999572f1661ee"
@@ -505,7 +512,6 @@ def test_event_with_attachment(
     relay = relay_with_processing()
     attachments_consumer = attachments_consumer()
     transactions_consumer = transactions_consumer()
-    outcomes_consumer = outcomes_consumer()
 
     # event attachments are always sent as chunks, and added to events
     envelope = Envelope(headers=[["event_id", event_id]])